# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Information System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: number;
  rollNumber: string;
  name: string;
  department: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  average?: number | null;
  /** @nullable */
  grade?: string | null;
  createdAt: string;
}

export interface StudentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  department: string;
  email?: string;
  phone?: string;
}

export interface StudentUpdate {
  name?: string;
  department?: string;
  email?: string;
  phone?: string;
}

export interface Course {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  credits?: number | null;
  /** @nullable */
  instructor?: string | null;
  createdAt: string;
}

export interface CourseInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  description?: string;
  credits?: number;
  instructor?: string;
}

export interface Enrollment {
  id: number;
  studentId: number;
  courseId: number;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  courseName?: string | null;
  /** @nullable */
  courseCode?: string | null;
  enrolledAt: string;
}

export interface EnrollmentInput {
  studentId: number;
  courseId: number;
}

export interface FeeRecord {
  id: number;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  tuitionFee: number;
  labFee: number;
  examFee: number;
  totalFee: number;
  paidAmount?: number;
  status: string;
  createdAt: string;
}

export interface FeeInput {
  studentId: number;
  tuitionFee: number;
  labFee: number;
  examFee: number;
  paidAmount?: number;
}

export type FeeUpdateStatus = typeof FeeUpdateStatus[keyof typeof FeeUpdateStatus];


export const FeeUpdateStatus = {
  paid: 'paid',
  unpaid: 'unpaid',
  partial: 'partial',
} as const;

export interface FeeUpdate {
  tuitionFee?: number;
  labFee?: number;
  examFee?: number;
  paidAmount?: number;
  status?: FeeUpdateStatus;
}

export interface PerformanceRecord {
  id: number;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  subject: string;
  marks: number;
  maxMarks: number;
  /** @nullable */
  percentage?: number | null;
  createdAt: string;
}

export interface PerformanceInput {
  studentId: number;
  /** @minLength 1 */
  subject: string;
  marks: number;
  maxMarks: number;
}

export interface PerformanceUpdate {
  subject?: string;
  marks?: number;
  maxMarks?: number;
}

export interface DashboardSummary {
  totalStudents: number;
  totalCourses: number;
  totalEnrollments: number;
  averageMarks: number;
  gradeACount: number;
  gradeBCount: number;
  gradeCCount: number;
  gradeDCount: number;
}

export interface GradeDistributionItem {
  grade: string;
  count: number;
}

export interface DepartmentStat {
  department: string;
  studentCount: number;
  averageMarks: number;
}

export interface FeeSummary {
  totalFees: number;
  totalPaid: number;
  totalOutstanding: number;
  paidCount: number;
  unpaidCount: number;
  partialCount: number;
}

export type ListStudentsParams = {
search?: string;
department?: string;
grade?: string;
sortBy?: ListStudentsSortBy;
sortOrder?: ListStudentsSortOrder;
};

export type ListStudentsSortBy = typeof ListStudentsSortBy[keyof typeof ListStudentsSortBy];


export const ListStudentsSortBy = {
  name: 'name',
  average: 'average',
  rollNumber: 'rollNumber',
  createdAt: 'createdAt',
} as const;

export type ListStudentsSortOrder = typeof ListStudentsSortOrder[keyof typeof ListStudentsSortOrder];


export const ListStudentsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ListEnrollmentsParams = {
studentId?: number;
courseId?: number;
};

export type ListFeesParams = {
studentId?: number;
status?: ListFeesStatus;
};

export type ListFeesStatus = typeof ListFeesStatus[keyof typeof ListFeesStatus];


export const ListFeesStatus = {
  paid: 'paid',
  unpaid: 'unpaid',
  partial: 'partial',
} as const;

export type ListPerformanceParams = {
studentId?: number;
};

export type GetTopStudentsParams = {
limit?: number;
};