# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Campus Information System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Course" }
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CourseInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Enrollments ───────────────────────────────────────────────────────────
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List all enrollments (optionally filtered by student or course)
      parameters:
        - name: studentId
          in: query
          schema: { type: integer }
        - name: courseId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Enrollment" }
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Enroll a student in a course
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnrollmentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Enrollment" }
  /enrollments/{id}:
    delete:
      operationId: deleteEnrollment
      tags: [enrollments]
      summary: Remove an enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Fees ──────────────────────────────────────────────────────────────────
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List all fee records
      parameters:
        - name: studentId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string, enum: [paid, unpaid, partial] }
      responses:
        "200":
          description: List of fee records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FeeRecord" }
    post:
      operationId: createFee
      tags: [fees]
      summary: Create a fee record for a student
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeeInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeeRecord" }
  /fees/{id}:
    put:
      operationId: updateFee
      tags: [fees]
      summary: Update a fee record (e.g. mark as paid)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeeUpdate" }
      responses:
        "200":
          description: Updated fee record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeeRecord" }

  # ── Performance ───────────────────────────────────────────────────────────
...
  /dashboard/top-students:
    get:
      operationId: getTopStudents
      tags: [dashboard]
      summary: Top performing students
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: Top students
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Student" }
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, rollNumber, name, department, createdAt]
      properties:
        id: { type: integer }
        rollNumber: { type: string }
        name: { type: string }
        department: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        average: { type: ["number", "null"] }
        grade: { type: ["string", "null"] }
        createdAt: { type: string }
...
    Course:
      type: object
      required: [id, name, code, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        description: { type: ["string", "null"] }
        credits: { type: ["integer", "null"] }
        instructor: { type: ["string", "null"] }
        createdAt: { type: string }
    CourseInput:
      type: object
      required: [name, code]
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1 }
        description: { type: string }
        credits: { type: integer }
        instructor: { type: string }

    # ── Enrollment ───────────────────────────────────────────────────────────
    Enrollment:
      type: object
      required: [id, studentId, courseId, enrolledAt]
      properties:
        id: { type: integer }
        studentId: { type: integer }
        courseId: { type: integer }
        studentName: { type: ["string", "null"] }
        courseName: { type: ["string", "null"] }
        courseCode: { type: ["string", "null"] }
        enrolledAt: { type: string }
    EnrollmentInput:
      type: object
      required: [studentId, courseId]
      properties:
        studentId: { type: integer }
        courseId: { type: integer }

    # ── Fee ──────────────────────────────────────────────────────────────────
    FeeRecord:
      type: object
      required: [id, studentId, tuitionFee, labFee, examFee, totalFee, status, createdAt]
      properties:
        id: { type: integer }
        studentId: { type: integer }
        studentName: { type: ["string", "null"] }
        tuitionFee: { type: number }
        labFee: { type: number }
        examFee: { type: number }
        totalFee: { type: number }
        paidAmount: { type: number }
        status: { type: string }
        createdAt: { type: string }
    FeeInput:
      type: object
      required: [studentId, tuitionFee, labFee, examFee]
      properties:
        studentId: { type: integer }
        tuitionFee: { type: number }
        labFee: { type: number }
        examFee: { type: number }
        paidAmount: { type: number }
    FeeUpdate:
      type: object
      properties:
        tuitionFee: { type: number }
        labFee: { type: number }
        examFee: { type: number }
        paidAmount: { type: number }
        status: { type: string, enum: [paid, unpaid, partial] }

    # ── Performance ───────────────────────────────────────────────────────────
...