# lib/api-zod/src/generated/types/feeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Information System API
 * OpenAPI spec version: 0.1.0
 */

export interface FeeInput {
  studentId: number;
  tuitionFee: number;
  labFee: number;
  examFee: number;
  paidAmount?: number;
}