# lib/api-zod/src/generated/types/feeRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Information System API
 * OpenAPI spec version: 0.1.0
 */

export interface FeeRecord {
  id: number;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  tuitionFee: number;
  labFee: number;
  examFee: number;
  totalFee: number;
  paidAmount?: number;
  status: string;
  createdAt: string;
}