# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Information System API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  rollNumber: string;
  name: string;
  department: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  average?: number | null;
  /** @nullable */
  grade?: string | null;
  createdAt: string;
}